pyproject.toml
[project]
name = "orchestrator-ai"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "golfergeek", email = "golfergeet@gmail.com"},
]
dependencies = []
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}

[project.group.dev.dependencies] # Correct way to specify dev dependencies for the root project
pytest = ">=8.2.0"
pytest-asyncio = ">=0.23.6"
pytest-mock = ">=3.12.0"


[project.optional-dependencies]
dev = [
    "pytest-mock>=3.14.0",
    "pytest-asyncio>=0.26.0",
    "httpx>=0.28.1",
    "fastapi>=0.115.12",
    "python-dotenv>=1.1.0",
    "openai>=1.78.1",
    "pydantic-settings>=2.9.1",
    "httpx-sse>=0.4.0",
]
[tool.pdm]
distribution = false

[tool.pytest.ini_options]
python_files = "tests.py test_*.py *_tests.py"
# Add pytest-mock to ensure mocker fixture is available
# Also add asyncio to ensure async tests are handled
addopts = "--asyncio-mode=auto"
# Set the default fixture loop scope to function to avoid deprecation warning
asyncio_default_fixture_loop_scope = "function"
